+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* `Contact` now has `is_emailer`, `is_interesting`, and `link_to_bio` props
 
 ### 1.8.7
 * More biographical information panels in chrono view